Adds support for path-based routing
Services that wish to be routed to by path
now just need to add a haproxy-pbr-name label.
This label should contain a regex for paths that
should route to the service.

The make target push-pbr creates and pushes an
image that will monitor pods with the haproxy-pbr-name
label. The monitor process writes config for haproxy
that routes to the service if the request path matches
the haproxy-pbr-name regex.