Adds default route to pbr haproxy
pbr haproxy now has default route to
wherever lookupCNAME(yetiapi.internal)
points. This means that if the incoming
request doesnt match any of the configured
path regexes, the request should be routed
to the old API.

amendment 1: added * after {SNI} in haproxy
config template. You cannot put '*' in an
rc so we'll just always do it. Haproxy
now matches paths that are haproxy-pbr-name*

amendment 2: now removes pods that dont match
the sni of an existing service even if their
IPs match

this commit might be squashable