Simplifies the config updating process
Rather than determine which pods have been
added or removed it now just determines if
a change has occurred.

If a change has occurred it rewrites the whole
config rather than just parts of it because
theres no reason not to.

Testing: the tests pass and its running in shared-yeti