fix pidfile logging bug
notifyHAProxy was logging a warning because it
was reading an empty pid file and couldn't find
the pid. NotifyHAProxy runs the correct command
when the pidfile is empty (it starts haproxy) but
it was still looking for an old pid to wait for.

Now it checks if the pidString is empty and if
that is the case it doesn't try to find a pid